almost there

+++ app/Http/Controllers/ProductController.php
@@ -30,8 +30,22 @@ class ProductController extends Controller
     public function destroy($id)
     {
         $product = Product::findOrFail($id);
+        if ($product->cartsProducts()->exists()) {
+            return redirect()->route('admin.products')->with('error', 'No se puede eliminar el producto porque está en un carrito de compra.');
+        }
+        try {
         $product->delete();
+
         return redirect()->route('admin.products')->with('success', 'Producto o servicio eliminado satisfactoriamente 😀');
+        } catch (QueryException $e) {
+            if ($e->errorInfo[1] == 1451) {
+                return redirect()->route('admin.products')->with('error', 'No se puede eliminar el producto porque está en un carrito de compra.');
+            }
+
+            return redirect()->route('admin.products')->with('error', 'Error al eliminar el producto: ' . $e->getMessage());
+        } catch (\Exception $e) {
+            return redirect()->route('admin.products')->with('error', 'Ocurrió un error inesperado: ' . $e->getMessage());
+        }
     }
 
     public function loadAddProducts()
